fix: remove asp dependencies from Offer.Library
create a new package with dependencies to asp which can be included by the services

Refs: CPLP-2344
