+++ src/Portal.Backend.sln
@@ -196,6 +196,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Processes.Library", "proces
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Processes.Library.Tests", "..\tests\processes\Processes.Library.Tests\Processes.Library.Tests.csproj", "{5E80DEEA-B254-425C-8220-27EEF47C10BD}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Offers.Library.Web", "marketplace\Offers.Library.Web\Offers.Library.Web.csproj", "{EC493B36-9E14-4CAF-973F-FB96FDAF546F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Offer.Library.Web.Tests", "..\tests\marketplace\Offer.Library.Web.Tests\Offer.Library.Web.Tests.csproj", "{1694E75F-ABCE-4573-B805-18ED50F148FD}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -1214,6 +1218,30 @@ Global
 		{5E80DEEA-B254-425C-8220-27EEF47C10BD}.Release|x64.Build.0 = Release|Any CPU
 		{5E80DEEA-B254-425C-8220-27EEF47C10BD}.Release|x86.ActiveCfg = Release|Any CPU
 		{5E80DEEA-B254-425C-8220-27EEF47C10BD}.Release|x86.Build.0 = Release|Any CPU
+		{EC493B36-9E14-4CAF-973F-FB96FDAF546F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EC493B36-9E14-4CAF-973F-FB96FDAF546F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EC493B36-9E14-4CAF-973F-FB96FDAF546F}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{EC493B36-9E14-4CAF-973F-FB96FDAF546F}.Debug|x64.Build.0 = Debug|Any CPU
+		{EC493B36-9E14-4CAF-973F-FB96FDAF546F}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{EC493B36-9E14-4CAF-973F-FB96FDAF546F}.Debug|x86.Build.0 = Debug|Any CPU
+		{EC493B36-9E14-4CAF-973F-FB96FDAF546F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EC493B36-9E14-4CAF-973F-FB96FDAF546F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EC493B36-9E14-4CAF-973F-FB96FDAF546F}.Release|x64.ActiveCfg = Release|Any CPU
+		{EC493B36-9E14-4CAF-973F-FB96FDAF546F}.Release|x64.Build.0 = Release|Any CPU
+		{EC493B36-9E14-4CAF-973F-FB96FDAF546F}.Release|x86.ActiveCfg = Release|Any CPU
+		{EC493B36-9E14-4CAF-973F-FB96FDAF546F}.Release|x86.Build.0 = Release|Any CPU
+		{1694E75F-ABCE-4573-B805-18ED50F148FD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1694E75F-ABCE-4573-B805-18ED50F148FD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1694E75F-ABCE-4573-B805-18ED50F148FD}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{1694E75F-ABCE-4573-B805-18ED50F148FD}.Debug|x64.Build.0 = Debug|Any CPU
+		{1694E75F-ABCE-4573-B805-18ED50F148FD}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{1694E75F-ABCE-4573-B805-18ED50F148FD}.Debug|x86.Build.0 = Debug|Any CPU
+		{1694E75F-ABCE-4573-B805-18ED50F148FD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1694E75F-ABCE-4573-B805-18ED50F148FD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1694E75F-ABCE-4573-B805-18ED50F148FD}.Release|x64.ActiveCfg = Release|Any CPU
+		{1694E75F-ABCE-4573-B805-18ED50F148FD}.Release|x64.Build.0 = Release|Any CPU
+		{1694E75F-ABCE-4573-B805-18ED50F148FD}.Release|x86.ActiveCfg = Release|Any CPU
+		{1694E75F-ABCE-4573-B805-18ED50F148FD}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -1306,5 +1334,7 @@ Global
 		{C356AA93-5BEE-44CD-A905-EFD325ED7578} = {323C198D-A8C6-4EB0-8B79-72624275E35F}
 		{CC063A63-8282-4293-9A0E-2598ADCA747B} = {323C198D-A8C6-4EB0-8B79-72624275E35F}
 		{58D5C0B4-3C5C-4BE4-82EA-4CB9940943B1} = {282CEF03-292F-4A49-83C6-997567D0FF5F}
+		{EC493B36-9E14-4CAF-973F-FB96FDAF546F} = {A878BDF1-6DB6-4BA5-A724-92885A710856}
+		{1694E75F-ABCE-4573-B805-18ED50F148FD} = {323C198D-A8C6-4EB0-8B79-72624275E35F}
 	EndGlobalSection
 EndGlobal
