+++ src/marketplace/Apps.Service/Apps.Service.csproj
@@ -41,6 +41,7 @@
     <ProjectReference Include="..\..\mailing\Mailing.SendMail\Mailing.SendMail.csproj" />
     <ProjectReference Include="..\..\portalbackend\PortalBackend.PortalEntities\PortalBackend.PortalEntities.csproj" />
     <ProjectReference Include="..\..\notifications\Notifications.Library\Notifications.Library.csproj" />
+    <ProjectReference Include="..\Offers.Library.Web\Offers.Library.Web.csproj" />
     <ProjectReference Include="..\Offers.Library\Offers.Library.csproj" />
     <ProjectReference Include="..\..\framework\Framework.BaseDependencies\Framework.BaseDependencies.csproj" />
   </ItemGroup>
