+++ src/marketplace/Offers.Library.Web/DependencyInjection/OfferSetupServiceExtensions.cs
@@ -1,0 +1,30 @@
+/********************************************************************************
+ * Copyright (c) 2021, 2023 BMW Group AG
+ * Copyright (c) 2021, 2023 Contributors to the Eclipse Foundation
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Apache License, Version 2.0 which is available at
+ * https://www.apache.org/licenses/LICENSE-2.0.
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ ********************************************************************************/
+
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Org.Eclipse.TractusX.Portal.Backend.Offers.Library.Web.DependencyInjection;
+
+public static class OfferSetupServiceCollectionExtension
+{
+    public static IServiceCollection AddOfferDocumentServices(this IServiceCollection services) =>
+        services
+            .AddTransient<IOfferDocumentService, OfferDocumentService>();
+}
