+++ src/marketplace/Offers.Library/Offers.Library.csproj
@@ -31,7 +31,6 @@
       <ProjectReference Include="..\..\framework\Framework.HttpClient\Framework.HttpClient.csproj" />
       <ProjectReference Include="..\..\framework\Framework.IO\Framework.IO.csproj" />
       <ProjectReference Include="..\..\framework\Framework.Logging\Framework.Logging.csproj" />
-      <ProjectReference Include="..\..\framework\Framework.Web\Framework.Web.csproj" />
       <ProjectReference Include="..\..\mailing\Mailing.SendMail\Mailing.SendMail.csproj" />
       <ProjectReference Include="..\..\portalbackend\PortalBackend.DBAccess\PortalBackend.DBAccess.csproj" />
       <ProjectReference Include="..\..\notifications\Notifications.Library\Notifications.Library.csproj" />
